Add option to supply additional local agent paths
This allows local agents, such as the one provided by 1Password to be
supplied on the command line.

Added some additional logic to detect if the local user is current
active based on keyboard/mouse activity and, if so, prefer that local
agent over any other that may be active (e.g a laptop that is also
shelled into the desktop machine, but you forgot to sign out of before
coming upstairs after having lunch.  Not that you would do that.)